style: 차트 그라데이션 커스텀
